server/requirements.txt
dependency-injector
environs
fastapi
google-auth[requests]
nanoid
python-multipart
pyjwt
psycopg2-binary
sqlalchemy
sqlmodel
uvicorn[standard]